requirements.txt
# --- PyTorch (nightly) ---
# Primary index is the official PyTorch nightly CUDA 12.8 (cu128) wheel repository.
# We keep PyPI as an extra index for all other scientific Python packages.
--pre
--index-url https://download.pytorch.org/whl/nightly/cu128
--extra-index-url https://pypi.org/simple

torch
torchvision
torchaudio

# --- Core numerics ---
numpy>=1.26
scipy>=1.11
pandas>=2.1

# --- Scientific IO / config / logging ---
h5py>=3.10
pyyaml>=6.0
tqdm>=4.66
joblib>=1.3
nvidia-ml-py>=12.560
datalad>=0.19

# --- EEG/BIDS ---
mne>=1.6
mne-bids>=0.15
pybids>=0.16
pymatreader>=1.1.0

# --- Modeling / stats ---
scikit-learn>=1.4
statsmodels>=0.14
gpytorch>=1.11

# --- Plotting / reports ---
matplotlib>=3.8
